+++ firmware/defmt-test/macros/src/lib.rs
@@ -47,22 +47,22 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                 let mut ignore = false;
 
                 f.attrs.retain(|attr| {
-                    if attr.path.is_ident("init") {
+                    if attr.path().is_ident("init") {
                         test_kind = Some(Attr::Init);
                         false
-                    } else if attr.path.is_ident("test") {
+                    } else if attr.path().is_ident("test") {
                         test_kind = Some(Attr::Test);
                         false
-                    } else if attr.path.is_ident("before_each") {
+                    } else if attr.path().is_ident("before_each") {
                         test_kind = Some(Attr::BeforeEach);
                         false
-                    } else if attr.path.is_ident("after_each") {
+                    } else if attr.path().is_ident("after_each") {
                         test_kind = Some(Attr::AfterEach);
                         false
-                    } else if attr.path.is_ident("should_error") {
+                    } else if attr.path().is_ident("should_error") {
                         should_error = true;
                         false
-                    } else if attr.path.is_ident("ignore") {
+                    } else if attr.path().is_ident("ignore") {
                         ignore = true;
                         false
                     } else {
@@ -520,7 +520,7 @@ fn extract_cfgs(attrs: &[Attribute]) -> Vec<Attribute> {
     let mut cfgs = vec![];
 
     for attr in attrs {
-        if attr.path.is_ident("cfg") {
+        if attr.path().is_ident("cfg") {
             cfgs.push(attr.clone());
         }
     }